test : session check filter 변경에 따른 업데이트
issue key : SB2-12
